requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2
pydantic-settings
python-jose[cryptography]
passlib
python-multipart
psycopg[binary]
